fork spring-projects

+++ spring-web/src/main/java/org/springframework/web/multipart/support/StandardServletMultipartResolver.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2002-2014 the original author or authors.
+ * Copyright 2002-2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -21,6 +21,7 @@ import javax.servlet.http.Part;
 
 import org.apache.commons.logging.LogFactory;
 
+import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartException;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
 import org.springframework.web.multipart.MultipartResolver;
@@ -40,8 +41,22 @@ import org.springframework.web.multipart.MultipartResolver;
  * storage locations need to be applied at that servlet registration level;
  * Servlet 3.0 does not allow for them to be set at the MultipartResolver level.
  *
+ * <pre class="code">
+ * public class AppInitializer extends AbstractAnnotationConfigDispatcherServletInitializer {
+ *	 // ...
+ *	 &#064;Override
+ *	 protected void customizeRegistration(ServletRegistration.Dynamic registration) {
+ *     // Optionally also set maxFileSize, maxRequestSize, fileSizeThreshold
+ *     registration.setMultipartConfig(new MultipartConfigElement("/tmp"));
+ *   }
+ * }
+ * </pre>
+ *
  * @author Juergen Hoeller
  * @since 3.1
+ * @see #setResolveLazily
+ * @see HttpServletRequest#getParts()
+ * @see org.springframework.web.multipart.commons.CommonsMultipartResolver
  */
 public class StandardServletMultipartResolver implements MultipartResolver {
 
@@ -55,6 +70,7 @@ public class StandardServletMultipartResolver implements MultipartResolver {
 	 * corresponding exceptions at the time of the {@link #resolveMultipart} call.
 	 * Switch this to "true" for lazy multipart parsing, throwing parse exceptions
 	 * once the application attempts to obtain multipart files or parameters.
+	 * @since 3.2.9
 	 */
 	public void setResolveLazily(boolean resolveLazily) {
 		this.resolveLazily = resolveLazily;
@@ -63,12 +79,7 @@ public class StandardServletMultipartResolver implements MultipartResolver {
 
 	@Override
 	public boolean isMultipart(HttpServletRequest request) {
-		// Same check as in Commons FileUpload...
-		if (!"post".equals(request.getMethod().toLowerCase())) {
-			return false;
-		}
-		String contentType = request.getContentType();
-		return (contentType != null && contentType.toLowerCase().startsWith("multipart/"));
+		return StringUtils.startsWithIgnoreCase(request.getContentType(), "multipart/");
 	}
 
 	@Override
@@ -78,17 +89,20 @@ public class StandardServletMultipartResolver implements MultipartResolver {
 
 	@Override
 	public void cleanupMultipart(MultipartHttpServletRequest request) {
-		// To be on the safe side: explicitly delete the parts,
-		// but only actual file parts (for Resin compatibility)
-		try {
-			for (Part part : request.getParts()) {
-				if (request.getFile(part.getName()) != null) {
-					part.delete();
+		if (!(request instanceof AbstractMultipartHttpServletRequest) ||
+				((AbstractMultipartHttpServletRequest) request).isResolved()) {
+			// To be on the safe side: explicitly delete the parts,
+			// but only actual file parts (for Resin compatibility)
+			try {
+				for (Part part : request.getParts()) {
+					if (request.getFile(part.getName()) != null) {
+						part.delete();
+					}
 				}
 			}
-		}
-		catch (Exception ex) {
-			LogFactory.getLog(getClass()).warn("Failed to perform cleanup of multipart items", ex);
+			catch (Throwable ex) {
+				LogFactory.getLog(getClass()).warn("Failed to perform cleanup of multipart items", ex);
+			}
 		}
 	}
